Do not show UI when the apply button is clicked
The plugin should only show additional dialog boxes after the user
clicks Ok.
This improves the user experience when "layering" multiple G'MIC
effects in a single session.

Changed the GmicPipeServer synchronization context to use Send instead
of Post.
Waiting for the call to complete matches the behavior when there is no
synchronization context, and fixes a potential race condition.