Add support for reading the document layers
Paint.NET 5.0 added the ability for plugins that use the new Effect
base classes to read all of the document layers.
The plugin has been converted to a BitmapEffect, which is the CPU-based
replacement for the classic Effect class.

The workaround of reading a second input image from the clipboard is
now obsolete and has been removed.